app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
Werkzeug==3.0.3
openpyxl==3.1.2
reportlab==4.1.0
python-dotenv==1.0.1
gunicorn==22.0.0
